static metrics calculation moved to separate file
